C# Coordinate Systems
Task form exam C# - Input data is being read from the console.
The number X is on the first input line.
The number Y is on the second input line.
The input data will always be valid and in the format described. There
is no need to check it explicitly.
Output
The output data must be printed on the console.
On the only output line you must print an integer number between 0 and
6, depending on the location of the given point in the coordinate
system.